+++ singularity-opac/neutrinos/mean_s_opacity_neutrinos.hpp
@@ -187,7 +187,8 @@ class MeanSOpacity {
               Real kappaPlanck = singularity_opac::robust::ratio(
                   kappaPlanckNum, kappaPlanckDenom);
               Real kappaRosseland =
-                  kappaPlanck > singularity_opac::robust::SMALL()
+                  (kappaPlanck > singularity_opac::robust::SMALL() &&
+                   kappaRosselandNum > singularity_opac::robust::SMALL())
                       ? singularity_opac::robust::ratio(kappaRosselandDenom,
                                                         kappaRosselandNum)
                       : 0.;
