+++ singularity-opac/photons/mean_opacity_photons.hpp
@@ -1,5 +1,5 @@
 // ======================================================================
-// © 2022-2024. Triad National Security, LLC. All rights reserved.  This
+// © 2026. Triad National Security, LLC. All rights reserved.  This
 // program was produced under U.S. Government contract
 // 89233218CNA000001 for Los Alamos National Laboratory (LANL), which
 // is operated by Triad National Security, LLC for the U.S.
@@ -16,20 +16,25 @@
 #ifndef SINGULARITY_OPAC_PHOTONS_MEAN_OPACITY_PHOTONS_
 #define SINGULARITY_OPAC_PHOTONS_MEAN_OPACITY_PHOTONS_
 
+// This file was made in part with generative AI.
+
+#include <cassert>
 #include <cmath>
-#include <filesystem>
-#include <fstream>
+#include <cstdio>
+#include <limits>
+#include <string>
+#include <vector>
 
 #include <ports-of-call/portability.hpp>
-#include <singularity-opac/base/radiation_types.hpp>
+#include <singularity-opac/base/opac_error.hpp>
 #include <singularity-opac/base/robust_utils.hpp>
 #include <singularity-opac/base/sp5.hpp>
 #include <singularity-opac/constants/constants.hpp>
-#include <spiner/databox.hpp>
-
+#include <singularity-opac/photons/mean_photon_types.hpp>
 #include <singularity-opac/photons/mean_photon_variant.hpp>
 #include <singularity-opac/photons/non_cgs_photons.hpp>
 #include <singularity-opac/photons/thermal_distributions_photons.hpp>
+#include <spiner/databox.hpp>
 
 namespace singularity {
 namespace photons {
@@ -44,66 +49,76 @@ template <typename pc = PhysicalConstantsCGS>
 class MeanOpacity {
  public:
   using PC = pc;
+  using DataBox = Spiner::DataBox<Real>;
 
   MeanOpacity() = default;
-  template <typename Opacity>
+
+  template <typename Opacity, typename GroupBoundsIndexer>
   MeanOpacity(const Opacity &opac, const Real lRhoMin, const Real lRhoMax,
-              const int NRho, const Real lTMin, const Real lTMax, const int NT,
-              Real *lambda = nullptr) {
-    MeanOpacityImpl_<Opacity, true>(opac, lRhoMin, lRhoMax, NRho, lTMin, lTMax,
-                                    NT, -1., -1., 100, lambda);
+                    const int NRho, const Real lTMin, const Real lTMax,
+                    const int NT, const GroupBoundsIndexer &group_bounds,
+                    const int ngroups, const int NNuPerGroup = 64,
+                    Real *lambda = nullptr) {
+    MeanOpacityImpl_(opac, lRhoMin, lRhoMax, NRho, lTMin, lTMax, NT,
+                           group_bounds, ngroups, NNuPerGroup, lambda);
   }
 
-  template <typename Opacity>
-  MeanOpacity(const Opacity &opac, const Real lRhoMin, const Real lRhoMax,
-              const int NRho, const Real lTMin, const Real lTMax, const int NT,
-              Real lNuMin, Real lNuMax, const int NNu, Real *lambda = nullptr) {
-    MeanOpacityImpl_<Opacity, false>(opac, lRhoMin, lRhoMax, NRho, lTMin, lTMax,
-                                     NT, lNuMin, lNuMax, NNu, lambda);
+  template <typename GroupBoundsIndexer>
+  MeanOpacity(const DataBox &kappaPlanck, const DataBox &kappaRosseland,
+                    const GroupBoundsIndexer &group_bounds) {
+    // Table-backed multigroup opacities always carry explicit group bounds.
+    // To represent [nu_max, infinity), the final bound must literally be
+    // IEEE +infinity, not a large finite proxy value.
+    LoadOpacityTables_(kappaPlanck, kappaRosseland, group_bounds);
   }
 
-  // construct Planck/Rosseland DataBox from ascii file
-  MeanOpacity(const std::string &filename) : filename_(filename.c_str()) {
-
-    // get number of density and temperature points
-    std::ifstream ff(filename.c_str());
-    const bool fexists = ff.good();
-
-    if (fexists) {
-
-      std::filesystem::path filePath(filename);
-      std::string extension = filePath.extension().string();
-
-      if (extension == ".txt") {
-        loadASCII(ff);
 #ifdef SPINER_USE_HDF
-      } else if (extension == ".hdf5" || extension == ".h5" ||
-                 extension == ".sp5") {
-        herr_t status = H5_SUCCESS;
-        hid_t file = H5Fopen(filename.c_str(), H5F_ACC_RDONLY, H5P_DEFAULT);
-        status += lkappa_.loadHDF(file, "Rosseland & Planck Mean Opacities");
-        status += H5Fclose(file);
+  MeanOpacity(const std::string &filename) {
+    // HDF-backed multigroup tables are expected to provide an ngroups + 1
+    // "group bounds" dataset. If the last group is [nu_max, infinity), then
+    // the final stored bound must be IEEE +infinity.
+    DataBox kappaPlanck;
+    DataBox kappaRosseland;
+    DataBox groupBounds;
+    herr_t status = H5_SUCCESS;
+    hid_t file = H5Fopen(filename.c_str(), H5F_ACC_RDONLY, H5P_DEFAULT);
+    status +=
+        kappaPlanck.loadHDF(file, SP5::MultigroupOpac::PlanckGroupOpacity);
+    status += kappaRosseland.loadHDF(
+        file, SP5::MultigroupOpac::RosselandGroupOpacity);
+    status += groupBounds.loadHDF(file, SP5::MultigroupOpac::GroupBounds);
+    status += H5Fclose(file);
 
-        if (status != H5_SUCCESS) {
-          OPAC_ERROR("photons::MeanOpacity: HDF5 error\n");
-        }
-#endif
-      } else {
-        OPAC_ERROR("photons::MeanOpacity: unrecognized file extension");
-      }
-
-    } else {
-      OPAC_ERROR("photons::MeanOpacity: file does not exist");
+    if (status != H5_SUCCESS) {
+      OPAC_ERROR("photons::MeanOpacity: HDF5 error\n");
     }
+
+    LoadOpacityTables_(kappaPlanck, kappaRosseland, groupBounds);
+    groupBounds.finalize();
+    kappaPlanck.finalize();
+    kappaRosseland.finalize();
   }
 
-#ifdef SPINER_USE_HDF
   void Save(const std::string &filename) const {
+    DataBox kappaPlanck;
+    DataBox kappaRosseland;
+    DataBox groupBounds;
+    ExportOpacityTables_(kappaPlanck, kappaRosseland);
+    ExportGroupBounds_(groupBounds);
+
     herr_t status = H5_SUCCESS;
     hid_t file =
         H5Fcreate(filename.c_str(), H5F_ACC_TRUNC, H5P_DEFAULT, H5P_DEFAULT);
-    status += lkappa_.saveHDF(file, "Rosseland & Planck Mean Opacities");
+    status +=
+        kappaPlanck.saveHDF(file, SP5::MultigroupOpac::PlanckGroupOpacity);
+    status += kappaRosseland.saveHDF(
+        file, SP5::MultigroupOpac::RosselandGroupOpacity);
+    status += groupBounds.saveHDF(file, SP5::MultigroupOpac::GroupBounds);
     status += H5Fclose(file);
+
+    kappaPlanck.finalize();
+    kappaRosseland.finalize();
+    groupBounds.finalize();
 
     if (status != H5_SUCCESS) {
       OPAC_ERROR("photons::MeanOpacity: HDF5 error\n");
@@ -111,17 +126,25 @@ class MeanOpacity {
   }
 #endif
 
-  PORTABLE_INLINE_FUNCTION void PrintParams() const {
-    printf("Mean opacity\n");
+  PORTABLE_INLINE_FUNCTION
+  void PrintParams() const {
+    printf("Photon multigroup opacity. ngroups = %d\n", ngroups_);
   }
 
   MeanOpacity GetOnDevice() {
     MeanOpacity other;
-    other.lkappa_ = Spiner::getOnDeviceDataBox(lkappa_);
+    other.lkappaPlanck_ = Spiner::getOnDeviceDataBox(lkappaPlanck_);
+    other.lkappaRosseland_ = Spiner::getOnDeviceDataBox(lkappaRosseland_);
+    other.groupBounds_ = Spiner::getOnDeviceDataBox(groupBounds_);
+    other.ngroups_ = ngroups_;
     return other;
   }
 
-  void Finalize() { lkappa_.finalize(); }
+  void Finalize() {
+    lkappaPlanck_.finalize();
+    lkappaRosseland_.finalize();
+    groupBounds_.finalize();
+  }
 
   PORTABLE_INLINE_FUNCTION RuntimePhysicalConstants
   GetRuntimePhysicalConstants() const {
@@ -129,196 +152,377 @@ class MeanOpacity {
   }
 
   PORTABLE_INLINE_FUNCTION
+  int ngroups() const noexcept { return ngroups_; }
+
+  PORTABLE_INLINE_FUNCTION
+  bool HasGroupBounds() const noexcept { return true; }
+
+  PORTABLE_INLINE_FUNCTION
   Real PlanckMeanAbsorptionCoefficient(const Real rho, const Real temp) const {
-    Real lRho = toLog_(rho);
-    Real lT = toLog_(temp);
-    return rho * fromLog_(lkappa_.interpToReal(lRho, lT, Planck));
+    return PlanckGroupAbsorptionCoefficient(rho, temp, 0);
   }
 
   PORTABLE_INLINE_FUNCTION
   Real RosselandMeanAbsorptionCoefficient(const Real rho,
                                           const Real temp) const {
-    Real lRho = toLog_(rho);
-    Real lT = toLog_(temp);
-    return rho * fromLog_(lkappa_.interpToReal(lRho, lT, Rosseland));
+    return RosselandGroupAbsorptionCoefficient(rho, temp, 0);
   }
 
-  // standard grey opacity functions
   PORTABLE_INLINE_FUNCTION
-  Real AbsorptionCoefficient(const Real rho, const Real temp,
+  Real PlanckGroupAbsorptionCoefficient(const Real rho, const Real temp,
+                                        const int group) const {
+    return GroupAbsorptionCoefficient_(lkappaPlanck_, rho, temp, group);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real RosselandGroupAbsorptionCoefficient(const Real rho, const Real temp,
+                                           const int group) const {
+    return GroupAbsorptionCoefficient_(lkappaRosseland_, rho, temp, group);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real AbsorptionCoefficient(const Real rho, const Real temp, const int group,
                              const int gmode = Rosseland) const {
-    Real lRho = toLog_(rho);
-    Real lT = toLog_(temp);
-    return rho * fromLog_(lkappa_.interpToReal(lRho, lT, gmode));
+    return (gmode == Planck)
+               ? PlanckGroupAbsorptionCoefficient(rho, temp, group)
+               : RosselandGroupAbsorptionCoefficient(rho, temp, group);
   }
 
   PORTABLE_INLINE_FUNCTION
   Real Emissivity(const Real rho, const Real temp, const int gmode = Rosseland,
                   Real *lambda = nullptr) const {
-    Real B = dist_.ThermalDistributionOfT(temp, lambda);
-    Real lRho = toLog_(rho);
-    Real lT = toLog_(temp);
-    return rho * fromLog_(lkappa_.interpToReal(lRho, lT, gmode)) * B;
+    if (ngroups_ != 1) {
+      OPAC_ERROR("photons::MeanOpacity: Emissivity only valid for ngroups==1");
+    }
+    PlanckDistribution<PC> dist;
+    Real B = dist.ThermalDistributionOfT(temp, lambda);
+    return AbsorptionCoefficient(rho, temp, 0, gmode) * B;
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  int GroupOfNu(const Real nu) const {
+    if (!(nu >= GroupBoundAt_(groupBounds_, 0) &&
+          nu <= GroupBoundAt_(groupBounds_, ngroups_))) {
+      OPAC_ERROR(
+          "photons::MeanOpacity: frequency is outside group bounds");
+    }
+    return GroupOfNuImpl_(nu);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real PlanckGroupAbsorptionCoefficientFromNu(const Real rho, const Real temp,
+                                              const Real nu) const {
+    return AbsorptionCoefficientFromNu(rho, temp, nu, Planck);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real RosselandGroupAbsorptionCoefficientFromNu(const Real rho,
+                                                 const Real temp,
+                                                 const Real nu) const {
+    return AbsorptionCoefficientFromNu(rho, temp, nu, Rosseland);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real AbsorptionCoefficientFromNu(const Real rho, const Real temp,
+                                   const Real nu,
+                                   const int gmode = Rosseland) const {
+    return AbsorptionCoefficient(rho, temp, GroupOfNu(nu), gmode);
   }
 
  private:
-  template <typename Opacity, bool AUTOFREQ>
+  PORTABLE_INLINE_FUNCTION
+  Real GroupAbsorptionCoefficient_(const DataBox &lkappa, const Real rho,
+                                   const Real temp, const int group) const {
+    const Real lRho = toLog_(rho);
+    const Real lT = toLog_(temp);
+    return rho * fromLog_(lkappa.interpToReal(lRho, lT, group));
+  }
+
+  template <typename GroupBoundsIndexer>
+  PORTABLE_INLINE_FUNCTION Real
+  GroupBoundAt_(const GroupBoundsIndexer &group_bounds, const int group) const {
+    return group_bounds[group];
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real GroupBoundAt_(const DataBox &group_bounds, const int group) const {
+    return group_bounds(group);
+  }
+
+  template <typename GroupBoundsIndexer>
+  void ValidateGroupBounds_(const GroupBoundsIndexer &group_bounds,
+                            const int ngroups) const {
+    if (ngroups <= 0) {
+      OPAC_ERROR("photons::MeanOpacity: ngroups must be positive");
+    }
+    for (int group = 0; group <= ngroups; ++group) {
+      const Real bound = GroupBoundAt_(group_bounds, group);
+      if (std::isnan(bound)) {
+        OPAC_ERROR("photons::MeanOpacity: group bounds must be finite "
+                   "or IEEE +infinity");
+      }
+      if (std::isinf(bound) && bound < 0.) {
+        OPAC_ERROR(
+            "photons::MeanOpacity: group bounds may not be -infinity");
+      }
+      if (group == 0) {
+        if (!(bound >= 0.)) {
+          OPAC_ERROR("photons::MeanOpacity: first group bound must be "
+                     "nonnegative");
+        }
+      } else if (!(bound > GroupBoundAt_(group_bounds, group - 1))) {
+        OPAC_ERROR("photons::MeanOpacity: group bounds must be strictly "
+                   "increasing");
+      }
+      if (!std::isfinite(bound) && group != ngroups) {
+        OPAC_ERROR(
+            "photons::MeanOpacity: only the final group bound may be "
+            "IEEE +infinity");
+      }
+    }
+  }
+
+  void ValidateOpacityTables_(const DataBox &kappaPlanck,
+                              const DataBox &kappaRosseland) const {
+    if (kappaPlanck.rank() != 3 || kappaRosseland.rank() != 3) {
+      OPAC_ERROR("photons::MeanOpacity: opacity tables must be rank 3");
+    }
+    for (int dim = 1; dim <= 3; ++dim) {
+      if (kappaPlanck.dim(dim) != kappaRosseland.dim(dim)) {
+        OPAC_ERROR("photons::MeanOpacity: table dimensions do not match");
+      }
+    }
+    if (kappaPlanck.dim(1) <= 0) {
+      OPAC_ERROR("photons::MeanOpacity: ngroups must be positive");
+    }
+    if (kappaPlanck.range(1) != kappaRosseland.range(1) ||
+        kappaPlanck.range(2) != kappaRosseland.range(2)) {
+      OPAC_ERROR("photons::MeanOpacity: table ranges do not match");
+    }
+  }
+
+  template <typename GroupBoundsIndexer>
+  void LoadOpacityTables_(const DataBox &kappaPlanck,
+                          const DataBox &kappaRosseland,
+                          const GroupBoundsIndexer &group_bounds) {
+    ValidateOpacityTables_(kappaPlanck, kappaRosseland);
+    ngroups_ = kappaPlanck.dim(1);
+    ValidateGroupBounds_(group_bounds, ngroups_);
+    SetGroupBounds_(group_bounds, ngroups_);
+    lkappaPlanck_.copyMetadata(kappaPlanck);
+    lkappaRosseland_.copyMetadata(kappaRosseland);
+    for (int i = 0; i < kappaPlanck.size(); ++i) {
+      lkappaPlanck_(i) = toLog_(kappaPlanck(i));
+      lkappaRosseland_(i) = toLog_(kappaRosseland(i));
+    }
+  }
+
+  void ExportOpacityTables_(DataBox &kappaPlanck,
+                            DataBox &kappaRosseland) const {
+    kappaPlanck.copyMetadata(lkappaPlanck_);
+    kappaRosseland.copyMetadata(lkappaRosseland_);
+    for (int i = 0; i < lkappaPlanck_.size(); ++i) {
+      kappaPlanck(i) = fromLog_(lkappaPlanck_(i));
+      kappaRosseland(i) = fromLog_(lkappaRosseland_(i));
+    }
+  }
+
+  void ExportGroupBounds_(DataBox &groupBounds) const {
+    groupBounds.resize(ngroups_ + 1);
+    for (int group = 0; group <= ngroups_; ++group) {
+      groupBounds(group) = groupBounds_(group);
+    }
+  }
+
+  template <typename GroupBoundsIndexer>
+  void SetGroupBounds_(const GroupBoundsIndexer &group_bounds,
+                       const int ngroups) {
+    groupBounds_.resize(ngroups + 1);
+    for (int group = 0; group <= ngroups; ++group) {
+      groupBounds_(group) = GroupBoundAt_(group_bounds, group);
+    }
+  }
+
+  template <typename SampleOp>
+  void ForEachGroupFrequencySample_(const Real temp, const Real nuMin,
+                                    const Real nuMax, const int NNuPerGroup,
+                                    SampleOp &&sample_op) const {
+    // For [0, ∞) or very wide ranges, use thermal-aware sampling
+    // For reasonable finite ranges, integrate over the full group bounds
+    const Real nu_thermal_min = 1.e-3 * PC::kb * temp / PC::h;
+    const Real nu_thermal_max = 1.e3 * PC::kb * temp / PC::h;
+
+    // Determine if we need special handling
+    const bool is_lower_extreme = (nuMin == 0.) || (nuMin < 0.1 * nu_thermal_min);
+    const bool is_upper_extreme = !std::isfinite(nuMax) || (nuMax > 10. * nu_thermal_max);
+
+    // Set integration bounds, but ensure they're valid
+    Real nu_sample_min = is_lower_extreme ? nu_thermal_min : nuMin;
+    Real nu_sample_max = is_upper_extreme ? nu_thermal_max : nuMax;
+
+    // If thermal-aware bounds are invalid, use a small but valid range within group bounds
+    if (nu_sample_min >= nu_sample_max) {
+      if (std::isfinite(nuMax) && nuMax > 0.) {
+        // Group is [0 or small, nuMax]: sample near nuMax
+        nu_sample_min = 0.5 * nuMax;
+        nu_sample_max = nuMax;
+      } else {
+        // Group extends to infinity: sample around thermal peak
+        nu_sample_min = 0.1 * nu_thermal_max;
+        nu_sample_max = nu_thermal_max;
+      }
+    }
+
+    // Use logarithmic spacing with midpoint rule
+    const Real lNuMin = toLog_(nu_sample_min);
+    const Real lNuMax = toLog_(nu_sample_max);
+    const Real dlnu = (lNuMax - lNuMin) / NNuPerGroup;
+    for (int inu = 0; inu < NNuPerGroup; ++inu) {
+      const Real lnu = lNuMin + (inu + 0.5) * dlnu;
+      const Real nu = fromLog_(lnu);
+      sample_op(nu, nu * dlnu);
+    }
+  }
+
+  void ThermalWeightsAtNu_(const PlanckDistribution<PC> &dist, const Real temp,
+                           const Real nu, Real &B, Real &dBdT) const {
+    const Real x = PC::h * nu / (PC::kb * temp);
+    if (x < 80.) {
+      B = dist.ThermalDistributionOfTNu(temp, nu);
+      dBdT = dist.DThermalDistributionOfTNuDT(temp, nu);
+      return;
+    }
+
+    const Real expMinusX = std::exp(-x);
+    B = (2. * PC::h * nu * nu * nu / (PC::c * PC::c)) * expMinusX;
+    dBdT = 2. * PC::h * PC::h * nu * nu * nu * nu * expMinusX /
+           (temp * temp * PC::c * PC::c * PC::kb);
+  }
+
+  template <typename Opacity, typename GroupBoundsIndexer>
   void MeanOpacityImpl_(const Opacity &opac, const Real lRhoMin,
-                        const Real lRhoMax, const int NRho, const Real lTMin,
-                        const Real lTMax, const int NT, Real lNuMin,
-                        Real lNuMax, const int NNu, Real *lambda = nullptr) {
+                              const Real lRhoMax, const int NRho,
+                              const Real lTMin, const Real lTMax, const int NT,
+                              const GroupBoundsIndexer &group_bounds,
+                              const int ngroups, const int NNuPerGroup,
+                              Real *lambda = nullptr) {
 #ifndef NDEBUG
     auto RPC = RuntimePhysicalConstants(PC());
     auto opc = opac.GetRuntimePhysicalConstants();
     assert(RPC == opc && "Physical constants are the same");
 #endif
 
-    lkappa_.resize(NRho, NT, 2);
-    lkappa_.setRange(1, lTMin, lTMax, NT);
-    lkappa_.setRange(2, lRhoMin, lRhoMax, NRho);
+    if (NNuPerGroup < 2) {
+      OPAC_ERROR("photons::MeanOpacity: NNuPerGroup must be at least 2");
+    }
+    ValidateGroupBounds_(group_bounds, ngroups);
 
-    // Fill tables
-    for (int iRho = 0; iRho < NRho; ++iRho) {
-      Real lRho = lkappa_.range(2).x(iRho);
-      Real rho = fromLog_(lRho);
-      for (int iT = 0; iT < NT; ++iT) {
-        Real lT = lkappa_.range(1).x(iT);
-        Real T = fromLog_(lT);
-        Real kappaPlanckNum = 0.;
-        Real kappaPlanckDenom = 0.;
-        Real kappaRosselandNum = 0.;
-        Real kappaRosselandDenom = 0.;
-        if (AUTOFREQ) {
-          lNuMin = toLog_(1.e-3 * PC::kb * fromLog_(lTMin) / PC::h);
-          lNuMax = toLog_(1.e3 * PC::kb * fromLog_(lTMax) / PC::h);
-        }
-        const Real dlnu = (lNuMax - lNuMin) / NNu;
-        // Integrate over frequency using midpoint rule
-        for (int inu = 0; inu < NNu; ++inu) {
-          const Real lnu = lNuMin + (inu + 0.5) * dlnu;
-          const Real nu = fromLog_(lnu);
-          const Real alpha = opac.AbsorptionCoefficient(rho, T, nu, lambda);
-          const Real B = opac.ThermalDistributionOfTNu(T, nu);
-          const Real dBdT = opac.DThermalDistributionOfTNuDT(T, nu);
-          kappaPlanckNum += alpha / rho * B * nu * dlnu;
-          kappaPlanckDenom += B * nu * dlnu;
+    ngroups_ = ngroups;
+    SetGroupBounds_(group_bounds, ngroups_);
+    lkappaPlanck_.resize(NRho, NT, ngroups_);
+    lkappaPlanck_.setRange(1, lTMin, lTMax, NT);
+    lkappaPlanck_.setRange(2, lRhoMin, lRhoMax, NRho);
+    lkappaRosseland_.copyMetadata(lkappaPlanck_);
 
-          if (alpha > singularity_opac::robust::SMALL()) {
-            kappaRosselandNum +=
-                singularity_opac::robust::ratio(rho, alpha) * dBdT * nu * dlnu;
-            kappaRosselandDenom += dBdT * nu * dlnu;
+    PlanckDistribution<PC> dist;
+    std::vector<Real> planckDenom(ngroups_, 0.);
+    std::vector<Real> rosselandDenom(ngroups_, 0.);
+
+    for (int iT = 0; iT < NT; ++iT) {
+      const Real lT = lkappaPlanck_.range(1).x(iT);
+      const Real T = fromLog_(lT);
+
+      for (int group = 0; group < ngroups_; ++group) {
+        Real Baccum = 0.;
+        Real dBdTaccum = 0.;
+        const Real nuMin = GroupBoundAt_(group_bounds, group);
+        const Real nuMax = GroupBoundAt_(group_bounds, group + 1);
+        ForEachGroupFrequencySample_(
+            T, nuMin, nuMax, NNuPerGroup, [&](const Real nu, const Real dnu) {
+              Real B = 0.;
+              Real dBdT = 0.;
+              ThermalWeightsAtNu_(dist, T, nu, B, dBdT);
+              Baccum += B * dnu;
+              dBdTaccum += dBdT * dnu;
+            });
+
+        planckDenom[group] = Baccum;
+        rosselandDenom[group] = dBdTaccum;
+      }
+
+      for (int iRho = 0; iRho < NRho; ++iRho) {
+        const Real lRho = lkappaPlanck_.range(2).x(iRho);
+        const Real rho = fromLog_(lRho);
+
+        for (int group = 0; group < ngroups_; ++group) {
+          Real kappaPlanckNum = 0.;
+          Real kappaRosselandNum = 0.;
+          const Real nuMin = GroupBoundAt_(group_bounds, group);
+          const Real nuMax = GroupBoundAt_(group_bounds, group + 1);
+          ForEachGroupFrequencySample_(
+              T, nuMin, nuMax, NNuPerGroup, [&](const Real nu, const Real dnu) {
+                const Real alpha =
+                    opac.AbsorptionCoefficient(rho, T, nu, lambda);
+                Real B = 0.;
+                Real dBdT = 0.;
+                ThermalWeightsAtNu_(dist, T, nu, B, dBdT);
+                kappaPlanckNum += alpha / rho * B * dnu;
+
+                if (alpha > singularity_opac::robust::SMALL()) {
+                  kappaRosselandNum +=
+                      singularity_opac::robust::ratio(rho, alpha) * dBdT * dnu;
+                }
+              });
+
+          const Real kappaPlanck = singularity_opac::robust::ratio(
+              kappaPlanckNum, planckDenom[group]);
+          const Real kappaRosseland =
+              (rosselandDenom[group] > singularity_opac::robust::SMALL() &&
+               kappaRosselandNum > singularity_opac::robust::SMALL())
+                  ? singularity_opac::robust::ratio(rosselandDenom[group],
+                                                    kappaRosselandNum)
+                  : 0.;
+
+          lkappaPlanck_(iRho, iT, group) = toLog_(kappaPlanck);
+          lkappaRosseland_(iRho, iT, group) = toLog_(kappaRosseland);
+          if (std::isnan(lkappaPlanck_(iRho, iT, group)) ||
+              std::isnan(lkappaRosseland_(iRho, iT, group))) {
+            OPAC_ERROR(
+                "photons::MeanOpacity: NAN in opacity evaluations");
           }
         }
-
-        Real kappaPlanck =
-            singularity_opac::robust::ratio(kappaPlanckNum, kappaPlanckDenom);
-        Real kappaRosseland = kappaPlanck > singularity_opac::robust::SMALL()
-                                  ? singularity_opac::robust::ratio(
-                                        kappaRosselandDenom, kappaRosselandNum)
-                                  : 0.;
-
-        lkappa_(iRho, iT, Rosseland) = toLog_(kappaRosseland);
-        lkappa_(iRho, iT, Planck) = toLog_(kappaPlanck);
-        if (std::isnan(lkappa_(iRho, iT, Rosseland)) ||
-            std::isnan(lkappa_(iRho, iT, Planck))) {
-          OPAC_ERROR("photons::MeanOpacity: NAN in opacity evaluations");
-        }
       }
     }
   }
 
-  // ASCII opacity file reader
-  void loadASCII(std::ifstream &ff) {
+  PORTABLE_INLINE_FUNCTION
+  Real toLog_(const Real x) const { return std::log10(std::abs(x) + EPS); }
 
-    int NRho = -1;
-    int NT = -1;
+  PORTABLE_INLINE_FUNCTION
+  Real fromLog_(const Real lx) const { return std::pow(10., lx); }
 
-    // line read from file
-    std::string fline;
-
-    // read 1st line of header to get sizes
-    std::getline(ff, fline);
-
-    // tokenize fline
-    char *cfline = const_cast<char *>(fline.c_str());
-    char *fl_tok = std::strtok(cfline, " ");
-
-    // move to next token to get number of density points
-    fl_tok = std::strtok(nullptr, " ");
-    NRho = std::stoi(fl_tok);
-
-    // move to next token to get number of temperature points
-    fl_tok = std::strtok(nullptr, " ");
-    fl_tok = std::strtok(nullptr, " ");
-    NT = std::stoi(fl_tok);
-
-    // read 2nd line of header to get min/max density
-    std::getline(ff, fline);
-    // tokenize fline
-    cfline = const_cast<char *>(fline.c_str());
-    fl_tok = std::strtok(cfline, " ");
-    fl_tok = std::strtok(nullptr, " ");
-    const Real RhoMin = std::stod(fl_tok);
-    fl_tok = std::strtok(nullptr, " ");
-    fl_tok = std::strtok(nullptr, " ");
-    const Real RhoMax = std::stod(fl_tok);
-
-    // read 3nd line of header to get min/max temperature
-    std::getline(ff, fline);
-    // tokenize fline
-    cfline = const_cast<char *>(fline.c_str());
-    fl_tok = std::strtok(cfline, " ");
-    fl_tok = std::strtok(nullptr, " ");
-    const Real TMin = std::stod(fl_tok);
-    fl_tok = std::strtok(nullptr, " ");
-    fl_tok = std::strtok(nullptr, " ");
-    const Real TMax = std::stod(fl_tok);
-
-    // reseize the Planck and Rosseland databox
-    lkappa_.resize(NRho, NT, 2);
-
-    // set rho-T ranges
-    const Real lTMin = toLog_(TMin);
-    const Real lTMax = toLog_(TMax);
-    const Real lRhoMin = toLog_(RhoMin);
-    const Real lRhoMax = toLog_(RhoMax);
-    lkappa_.setRange(1, lTMin, lTMax, NT);
-    lkappa_.setRange(2, lRhoMin, lRhoMax, NRho);
-
-    // fill tables
-    for (int iRho = 0; iRho < NRho; ++iRho) {
-      const Real lRho_i = lkappa_.range(2).x(iRho);
-      for (int iT = 0; iT < NT; ++iT) {
-
-        // get new file-line
-        std::getline(ff, fline);
-        cfline = const_cast<char *>(fline.c_str());
-        fl_tok = std::strtok(cfline, " ");
-
-        // populate Rosseland opacity [cm^2/g]
-        lkappa_(iRho, iT, Rosseland) = toLog_(std::stod(fl_tok));
-
-        // populate Planck opacity [cm^2/g]
-        fl_tok = std::strtok(nullptr, " ");
-        lkappa_(iRho, iT, Planck) = toLog_(std::stod(fl_tok));
-
-        if (std::isnan(lkappa_(iRho, iT, Rosseland)) ||
-            std::isnan(lkappa_(iRho, iT, Planck))) {
-          OPAC_ERROR("photons::MeanOpacity: NAN in parsed ASCII opacity");
-        }
+  PORTABLE_INLINE_FUNCTION
+  int GroupOfNuImpl_(const Real nu) const {
+    if (nu == GroupBoundAt_(groupBounds_, ngroups_)) {
+      return ngroups_ - 1;
+    }
+    int lower = 0;
+    int upper = ngroups_;
+    while (upper - lower > 1) {
+      const int middle = (lower + upper) / 2;
+      if (nu < GroupBoundAt_(groupBounds_, middle)) {
+        upper = middle;
+      } else {
+        lower = middle;
       }
     }
+    return lower;
   }
 
-  PORTABLE_INLINE_FUNCTION Real toLog_(const Real x) const {
-    return std::log10(std::abs(x) + EPS);
-  }
-  PORTABLE_INLINE_FUNCTION Real fromLog_(const Real lx) const {
-    return std::pow(10., lx);
-  }
-  Spiner::DataBox<Real> lkappa_;
-  const char *filename_;
-  PlanckDistribution<PC> dist_;
+  DataBox lkappaPlanck_;
+  DataBox lkappaRosseland_;
+  DataBox groupBounds_;
+  int ngroups_ = 0;
 };
 
 #undef EPS
@@ -327,9 +531,10 @@ class MeanOpacity {
 
 using MeanOpacityBase = impl::MeanOpacity<PhysicalConstantsCGS>;
 using MeanOpacity =
-    impl::MeanVariant<MeanOpacityBase, MeanNonCGSUnits<MeanOpacityBase>>;
+    impl::MeanVariant<MeanOpacityBase,
+                            MeanNonCGSUnits<MeanOpacityBase>>;
 
 } // namespace photons
 } // namespace singularity
 
-#endif // SINGULARITY_OPAC_PHOTONS_MEAN_OPACITY_PHOTONS__
+#endif // SINGULARITY_OPAC_PHOTONS_MEAN_OPACITY_PHOTONS_
