+++ singularity-opac/photons/mean_photon_s_variant.hpp
@@ -88,6 +88,70 @@ class MeanSVariant {
         s_opac_);
   }
 
+  PORTABLE_INLINE_FUNCTION
+  int ngroups() const noexcept {
+    return PortsOfCall::visit([](const auto &s_opac) { return s_opac.ngroups(); },
+                              s_opac_);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  bool HasGroupBounds() const noexcept {
+    return PortsOfCall::visit(
+        [](const auto &s_opac) { return s_opac.HasGroupBounds(); }, s_opac_);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real PlanckGroupScatteringCoefficient(const Real rho, const Real temp,
+                                        const int group) const {
+    return ScatteringCoefficient(rho, temp, group, Planck);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real RosselandGroupScatteringCoefficient(const Real rho, const Real temp,
+                                           const int group) const {
+    return ScatteringCoefficient(rho, temp, group, Rosseland);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real ScatteringCoefficient(const Real rho, const Real temp, const int group,
+                             const int gmode = Rosseland) const {
+    return PortsOfCall::visit(
+        [=](const auto &s_opac) {
+          return s_opac.ScatteringCoefficient(rho, temp, group, gmode);
+        },
+        s_opac_);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  int GroupOfNu(const Real nu) const {
+    return PortsOfCall::visit(
+        [=](const auto &s_opac) { return s_opac.GroupOfNu(nu); }, s_opac_);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real PlanckGroupScatteringCoefficientFromNu(const Real rho, const Real temp,
+                                              const Real nu) const {
+    return ScatteringCoefficientFromNu(rho, temp, nu, Planck);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real RosselandGroupScatteringCoefficientFromNu(const Real rho,
+                                                 const Real temp,
+                                                 const Real nu) const {
+    return ScatteringCoefficientFromNu(rho, temp, nu, Rosseland);
+  }
+
+  PORTABLE_INLINE_FUNCTION
+  Real ScatteringCoefficientFromNu(const Real rho, const Real temp,
+                                   const Real nu,
+                                   const int gmode = Rosseland) const {
+    return PortsOfCall::visit(
+        [=](const auto &s_opac) {
+          return s_opac.ScatteringCoefficientFromNu(rho, temp, nu, gmode);
+        },
+        s_opac_);
+  }
+
   inline void Finalize() noexcept {
     return PortsOfCall::visit([](auto &s_opac) { return s_opac.Finalize(); },
                               s_opac_);
