--- singularity-opac/photons/multigroup_photon_variant.hpp
@@ -1,162 +1,0 @@
-// ======================================================================
-// © 2026. Triad National Security, LLC. All rights reserved.  This
-// program was produced under U.S. Government contract
-// 89233218CNA000001 for Los Alamos National Laboratory (LANL), which
-// is operated by Triad National Security, LLC for the U.S.
-// Department of Energy/National Nuclear Security Administration. All
-// rights in the program are reserved by Triad National Security, LLC,
-// and the U.S. Department of Energy/National Nuclear Security
-// Administration. The Government is granted for itself and others
-// acting on its behalf a nonexclusive, paid-up, irrevocable worldwide
-// license in this material to reproduce, prepare derivative works,
-// distribute copies to the public, perform publicly and display
-// publicly, and to permit others to do so.
-// ======================================================================
-
-#ifndef SINGULARITY_OPAC_PHOTONS_MULTIGROUP_PHOTON_VARIANT_
-#define SINGULARITY_OPAC_PHOTONS_MULTIGROUP_PHOTON_VARIANT_
-
-// This file was made in part with generative AI.
-
-#include <string>
-#include <utility>
-
-#include <ports-of-call/portability.hpp>
-#include <ports-of-call/variant.hpp>
-#include <singularity-opac/photons/mean_photon_types.hpp>
-#include <singularity-opac/photons/photon_variant.hpp>
-
-namespace singularity {
-namespace photons {
-namespace impl {
-
-template <typename... Opacs>
-class MultigroupVariant {
- private:
-  opac_variant<Opacs...> opac_;
-
- public:
-  template <typename Choice,
-            typename std::enable_if<
-                !std::is_same<MultigroupVariant,
-                              typename std::decay<Choice>::type>::value,
-                bool>::type = true>
-  PORTABLE_FUNCTION MultigroupVariant(Choice &&choice)
-      : opac_(std::forward<Choice>(choice)) {}
-
-  MultigroupVariant() = default;
-
-  template <typename Choice,
-            typename std::enable_if<
-                !std::is_same<MultigroupVariant,
-                              typename std::decay<Choice>::type>::value,
-                bool>::type = true>
-  PORTABLE_FUNCTION MultigroupVariant &operator=(Choice &&opac) {
-    opac_ = std::forward<Choice>(opac);
-    return *this;
-  }
-
-  template <typename Choice,
-            typename std::enable_if<
-                !std::is_same<MultigroupVariant,
-                              typename std::decay<Choice>::type>::value,
-                bool>::type = true>
-  Choice get() {
-    return PortsOfCall::get<Choice>(opac_);
-  }
-
-  MultigroupVariant GetOnDevice() {
-    return PortsOfCall::visit(
-        [](auto &opac) { return opac_variant<Opacs...>(opac.GetOnDevice()); },
-        opac_);
-  }
-
-  PORTABLE_INLINE_FUNCTION RuntimePhysicalConstants
-  GetRuntimePhysicalConstants() const {
-    return PortsOfCall::visit(
-        [](const auto &opac) { return opac.GetRuntimePhysicalConstants(); },
-        opac_);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  int ngroups() const noexcept {
-    return PortsOfCall::visit([](const auto &opac) { return opac.ngroups(); },
-                              opac_);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  bool HasGroupBounds() const noexcept {
-    return PortsOfCall::visit(
-        [](const auto &opac) { return opac.HasGroupBounds(); }, opac_);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real PlanckGroupAbsorptionCoefficient(const Real rho, const Real temp,
-                                        const int group) const {
-    return AbsorptionCoefficient(rho, temp, group, Planck);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real RosselandGroupAbsorptionCoefficient(const Real rho, const Real temp,
-                                           const int group) const {
-    return AbsorptionCoefficient(rho, temp, group, Rosseland);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real AbsorptionCoefficient(const Real rho, const Real temp, const int group,
-                             const int gmode = Rosseland) const {
-    return PortsOfCall::visit(
-        [=](const auto &opac) {
-          return opac.AbsorptionCoefficient(rho, temp, group, gmode);
-        },
-        opac_);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  int GroupOfNu(const Real nu) const {
-    return PortsOfCall::visit(
-        [=](const auto &opac) { return opac.GroupOfNu(nu); }, opac_);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real PlanckGroupAbsorptionCoefficientFromNu(const Real rho, const Real temp,
-                                              const Real nu) const {
-    return AbsorptionCoefficientFromNu(rho, temp, nu, Planck);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real RosselandGroupAbsorptionCoefficientFromNu(const Real rho,
-                                                 const Real temp,
-                                                 const Real nu) const {
-    return AbsorptionCoefficientFromNu(rho, temp, nu, Rosseland);
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real AbsorptionCoefficientFromNu(const Real rho, const Real temp,
-                                   const Real nu,
-                                   const int gmode = Rosseland) const {
-    return PortsOfCall::visit(
-        [=](const auto &opac) {
-          return opac.AbsorptionCoefficientFromNu(rho, temp, nu, gmode);
-        },
-        opac_);
-  }
-
-  inline void Finalize() noexcept {
-    return PortsOfCall::visit([](auto &opac) { return opac.Finalize(); },
-                              opac_);
-  }
-
-#ifdef SPINER_USE_HDF
-  void Save(const std::string &filename) const {
-    return PortsOfCall::visit([=](const auto &opac) { opac.Save(filename); },
-                              opac_);
-  }
-#endif
-};
-
-} // namespace impl
-} // namespace photons
-} // namespace singularity
-
-#endif // SINGULARITY_OPAC_PHOTONS_MULTIGROUP_PHOTON_VARIANT_
