+++ singularity-opac/photons/non_cgs_photons.hpp
@@ -245,98 +245,17 @@ class MeanNonCGSUnits {
   using PC = typename MeanOpac::PC;
 
   MeanNonCGSUnits() = default;
-  MeanNonCGSUnits(MeanOpac &&mean_opac, const Real time_unit,
-                  const Real mass_unit, const Real length_unit,
-                  const Real temp_unit)
-      : mean_opac_(std::forward<MeanOpac>(mean_opac)), time_unit_(time_unit),
-        mass_unit_(mass_unit), length_unit_(length_unit), temp_unit_(temp_unit),
-        rho_unit_(mass_unit_ / (length_unit_ * length_unit_ * length_unit_)),
-        inv_emiss_unit_(length_unit_ * time_unit_ * time_unit_ * time_unit_ /
-                        mass_unit_) {}
-
-  auto GetOnDevice() {
-    return MeanNonCGSUnits<MeanOpac>(mean_opac_.GetOnDevice(), time_unit_,
-                                     mass_unit_, length_unit_, temp_unit_);
-  }
-  inline void Finalize() noexcept { mean_opac_.Finalize(); }
-
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return mean_opac_.nlambda(); }
-
-#ifdef SPINER_USE_HDF
-  void Save(const std::string &filename) const {
-    return mean_opac_.Save(filename);
-  }
-#endif
-
-  PORTABLE_INLINE_FUNCTION
-  Real PlanckMeanAbsorptionCoefficient(const Real rho, const Real temp) const {
-    const Real alpha = mean_opac_.PlanckMeanAbsorptionCoefficient(
-        rho_unit_ * rho, temp_unit_ * temp);
-    // alpha output in units of 1/cm. Want to convert out of CGS.
-    // multiplication by length_unit converts length to cm.
-    // division converts length from cm to unit system.
-    // thus multiplication converts (1/cm) to unit system.
-    return alpha * length_unit_;
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real RosselandMeanAbsorptionCoefficient(const Real rho,
-                                          const Real temp) const {
-    const Real alpha = mean_opac_.RosselandMeanAbsorptionCoefficient(
-        rho_unit_ * rho, temp_unit_ * temp);
-    // alpha output in units of 1/cm. Want to convert out of CGS.
-    // multiplication by length_unit converts length to cm.
-    // division converts length from cm to unit system.
-    // thus multiplication converts (1/cm) to unit system.
-    return alpha * length_unit_;
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real AbsorptionCoefficient(const Real rho, const Real temp,
-                             const int gmode = Rosseland) const {
-    const Real alpha = mean_opac_.AbsorptionCoefficient(
-        rho_unit_ * rho, temp_unit_ * temp, gmode);
-    return alpha * length_unit_;
-  }
-
-  PORTABLE_INLINE_FUNCTION
-  Real Emissivity(const Real rho, const Real temp, const int gmode = Rosseland,
-                  Real *lambda = nullptr) const {
-    const Real J =
-        mean_opac_.Emissivity(rho_unit_ * rho, temp_unit_ * temp, gmode);
-    return J * inv_emiss_unit_;
-  }
-
-  PORTABLE_INLINE_FUNCTION RuntimePhysicalConstants
-  GetRuntimePhysicalConstants() const {
-    return RuntimePhysicalConstants(PhysicalConstantsCGS(), time_unit_,
-                                    mass_unit_, length_unit_, temp_unit_);
-  }
-
- private:
-  MeanOpac mean_opac_;
-  Real time_unit_, mass_unit_, length_unit_, temp_unit_;
-  Real rho_unit_, inv_emiss_unit_;
-};
-
-template <typename MultigroupOpac>
-class MultigroupNonCGSUnits {
- public:
-  using PC = typename MultigroupOpac::PC;
-
-  MultigroupNonCGSUnits() = default;
-  MultigroupNonCGSUnits(MultigroupOpac &&multigroup_opac, const Real time_unit,
+  MeanNonCGSUnits(MeanOpac &&multigroup_opac, const Real time_unit,
                         const Real mass_unit, const Real length_unit,
                         const Real temp_unit)
-      : multigroup_opac_(std::forward<MultigroupOpac>(multigroup_opac)),
+      : multigroup_opac_(std::forward<MeanOpac>(multigroup_opac)),
         time_unit_(time_unit), mass_unit_(mass_unit), length_unit_(length_unit),
         temp_unit_(temp_unit),
         rho_unit_(mass_unit_ / (length_unit_ * length_unit_ * length_unit_)),
         freq_unit_(1. / time_unit_) {}
 
   auto GetOnDevice() {
-    return MultigroupNonCGSUnits<MultigroupOpac>(multigroup_opac_.GetOnDevice(),
+    return MeanNonCGSUnits<MeanOpac>(multigroup_opac_.GetOnDevice(),
                                                  time_unit_, mass_unit_,
                                                  length_unit_, temp_unit_);
   }
@@ -348,6 +267,11 @@ class MultigroupNonCGSUnits {
   PORTABLE_INLINE_FUNCTION
   bool HasGroupBounds() const noexcept {
     return multigroup_opac_.HasGroupBounds();
+  }
+
+  PORTABLE_INLINE_FUNCTION RuntimePhysicalConstants
+  GetRuntimePhysicalConstants() const {
+    return multigroup_opac_.GetRuntimePhysicalConstants();
   }
 
 #ifdef SPINER_USE_HDF
@@ -403,14 +327,8 @@ class MultigroupNonCGSUnits {
     return alpha * length_unit_;
   }
 
-  PORTABLE_INLINE_FUNCTION RuntimePhysicalConstants
-  GetRuntimePhysicalConstants() const {
-    return RuntimePhysicalConstants(PhysicalConstantsCGS(), time_unit_,
-                                    mass_unit_, length_unit_, temp_unit_);
-  }
-
  private:
-  MultigroupOpac multigroup_opac_;
+  MeanOpac multigroup_opac_;
   Real time_unit_, mass_unit_, length_unit_, temp_unit_;
   Real rho_unit_, freq_unit_;
 };
